Refactor and split code, favour single responsibility
